refactor(ui): scroll the transcript in (segment, row)
`scroll_top` was a document row count, meaningless without a width, so
every resize had to re-derive it: an anchor segment, a re-pin after the
reflow and a convergence argument that the reflow window would settle.

The scroll position is now a place in the document, `{ seg, row }`, which
no width can move. The anchor, `anchor_at`, `anchor_offset` and the re-pin
go away; render, clicks and the reflow window all walk from `scroll.seg`.
One `Layout` owns every row walk over the cached segments plus the
streaming tail, so the two are counted once and in one place.

Document rows survive only where they are the contract: the scrollbar and
`winsaveview`, both now `u32`, so a transcript past 65535 rows stops
clamping and clicking on the wrong segment.

+++ maki-lua/src/api/fn.rs
@@ -589,7 +589,7 @@ fn winrestview(
     view: Table,
 ) -> LuaResult<Pair<bool>> {
     let topline = view.get::<Option<i64>>("topline")?.unwrap_or(1);
-    let scroll_top = topline.saturating_sub(1).clamp(0, u16::MAX as i64) as u16;
+    let scroll_top = topline.saturating_sub(1).clamp(0, i64::from(u32::MAX)) as u32;
     try_pair!(ui_send(tx.as_ref(), UiAction::WinRestView { scroll_top }));
     Ok((Some(true), None))
 }
@@ -904,8 +904,8 @@ mod tests {
 
     #[test]
     fn winsaveview_reports_the_viewport_one_based() {
-        const SCROLL_TOP: u16 = 6;
-        const LINE_COUNT: u16 = 100;
+        const SCROLL_TOP: u32 = 6;
+        const LINE_COUNT: u32 = 100;
         const HEIGHT: u16 = 24;
 
         let (tx, rx) = flume::unbounded::<UiAction>();
@@ -926,8 +926,8 @@ mod tests {
         let (view, err): (Table, Option<String>) =
             smol::block_on(lua.load("return f.winsaveview()").eval_async()).unwrap();
         assert_eq!(err, None);
-        assert_eq!(view.get::<u16>("topline").unwrap(), SCROLL_TOP + 1);
-        assert_eq!(view.get::<u16>("line_count").unwrap(), LINE_COUNT);
+        assert_eq!(view.get::<u32>("topline").unwrap(), SCROLL_TOP + 1);
+        assert_eq!(view.get::<u32>("line_count").unwrap(), LINE_COUNT);
         assert_eq!(view.get::<u16>("height").unwrap(), HEIGHT);
         assert!(!view.get::<bool>("auto_scroll").unwrap());
     }
@@ -935,7 +935,7 @@ mod tests {
     #[test_case::test_case("{ topline = 12 }", 11 ; "explicit_topline")]
     #[test_case::test_case("{}", 0 ; "missing_topline_defaults_to_first_line")]
     #[test_case::test_case("{ topline = -5 }", 0 ; "below_range_clamps_to_first_line")]
-    fn winrestview_forwards_zero_based_scroll_top(arg: &str, expected: u16) {
+    fn winrestview_forwards_zero_based_scroll_top(arg: &str, expected: u32) {
         let (tx, rx) = flume::unbounded::<UiAction>();
         let lua = lua_with_view(Some(tx));
         let (ok, err): (bool, Option<String>) = smol::block_on(
