+++ maki-ui/src/components/btw_modal.rs
@@ -158,7 +158,7 @@ impl BtwModal {
         frame.render_widget(paragraph, padded);
 
         if total > viewport_h {
-            render_vertical_scrollbar(frame, inner, total, scroll);
+            render_vertical_scrollbar(frame, inner, u32::from(total), u32::from(scroll));
         }
 
         popup
