+++ maki-ui/src/components/input.rs
@@ -430,7 +430,7 @@ impl InputBox {
 
         if max_scroll > 0 {
             let inner = area.inner(ratatui::layout::Margin::new(0, 1));
-            render_vertical_scrollbar(frame, inner, total_vl, self.scroll_y);
+            render_vertical_scrollbar(frame, inner, u32::from(total_vl), u32::from(self.scroll_y));
         }
     }
